Move ort session handler to a separate header
